Move edit files to own directory
