+++ src/directory/index.js
@@ -18,7 +18,7 @@ import metadata from './block.json';
  */
 registerBlockType( metadata.name, {
 	/**
-	 * @see ./edit.js
+	 * @see ./edit/index.js
 	 */
 	edit: Edit,
 } );
